+++ extensions/world-clock/src/types/preferences.ts
@@ -1,4 +1,6 @@
 export interface Preferences {
+  itemLayout: string;
+  itemSize: string;
   rememberTag: boolean;
   hour24: boolean;
 }